+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 5.1.0
+- Updated internal libraries.
+- Bumped minimum Dart SDK to 2.17.
+- Added `MultipartFileRecreatablep documentation.
+- Refactors static constructors to factories.
+- Adds a new `MultipartFileRecreatable.fromBytes` factory compatible with web.
+- Added a new `headers` parameter.
+- You can now read the file's content with `MultipartFileRecreatable.data`.
+- **Breaking:** Renamed `filename` to `fileName`.
+- **Breaking:** `fileName` is now a named parameter.
+- **Breaking:** Removed `MultipartFileRecreatable.filePath`.
+
 ## 5.0.0
 - Add supporting of the new dio 5.+